Move compression function to internal library
